提交文件

+++ src/main/java/com/mzhotel/sm/userInfo/service/impl/UserInfoServiceImpl.java
@@ -1,5 +1,6 @@
 package com.mzhotel.sm.userInfo.service.impl;
 
+import com.aliyuncs.utils.StringUtils;
 import com.mzhotel.sm.common.ContextHolderUtils;
 import com.mzhotel.sm.login.dto.QueryUserInfo;
 import com.mzhotel.sm.pageUtil.PageResult;
@@ -17,7 +18,7 @@ import java.util.Date;
 import java.util.List;
 
 @Service
-public class UserInfoServiceImpl implements UserInfoService{
+public class UserInfoServiceImpl implements UserInfoService {
 
     @Autowired
     UserInfoMapper userInfoMapper;
@@ -27,63 +28,67 @@ public class UserInfoServiceImpl implements UserInfoService{
 
     public String getCurrUser() {
         HttpSession session = ContextHolderUtils.getSession();
-        return (String)session.getAttribute("userId");
+        return (String) session.getAttribute("userId");
     }
 
     @Override
-    public UserInfo getUserInfoByUserName(QueryUserInfo queryUserInfo){
+    public UserInfo getUserInfoByUserName(QueryUserInfo queryUserInfo) {
         return userInfoMapper.getUserInfoByUserName(queryUserInfo);
     }
 
     @Override
-    public List<Role> getRoleInfo(QueryUserInfo queryUserInfo){
+    public List<Role> getRoleInfo(QueryUserInfo queryUserInfo) {
         return userInfoMapper.getRoleInfo(queryUserInfo);
     }
 
     @Override
-    public List<UserInfo> getUserInfo(QueryUserInfo queryUserInfo){
+    public List<UserInfo> getUserInfo(QueryUserInfo queryUserInfo) {
         return userInfoMapper.getUserInfo(queryUserInfo);
     }
 
     @Override
-    public PageResult<UserInfo> queryPage(QueryUserInfo queryUserInfo){
-        return myBatisDAO.selectPage(getUserInfo(queryUserInfo),queryUserInfo.getPageNum(),queryUserInfo.getPageSize());
+    public PageResult<UserInfo> queryPage(QueryUserInfo queryUserInfo) {
+        return myBatisDAO.selectPage(getUserInfo(queryUserInfo), queryUserInfo.getPageNum(), queryUserInfo.getPageSize());
     }
 
     @Transactional
     @Override
-    public int deleteByPrimaryKey(String id) {
-        return userInfoMapper.deleteByPrimaryKey(id);
+    public void delete(String id) {
+        if (StringUtils.isEmpty(id)) {
+            return;
+        }
+        userInfoMapper.deleteByPrimaryKey(id);
     }
 
     @Transactional
     @Override
-    public int insert(UserInfo record) {
+    public UserInfo insert(UserInfo record) {
         record.setCreatedBy(getCurrUser());
         record.setCreatedDate(new Date());
-        return userInfoMapper.insert(record);
-    }
-
-    @Transactional
-    @Override
-    public int insertSelective(UserInfo record) {
-        return userInfoMapper.insertSelective(record);
+        int result = userInfoMapper.insert(record);
+        if (result == 1) {
+            return selectOne(record.getId());
+        }
+        return null;
     }
 
     @Override
-    public UserInfo selectByPrimaryKey(String id) {
+    public UserInfo selectOne(String id) {
+        if (StringUtils.isEmpty(id)) {
+            return null;
+        }
         return userInfoMapper.selectByPrimaryKey(id);
     }
 
     @Transactional
     @Override
-    public int updateByPrimaryKeySelective(UserInfo record) {
-        return userInfoMapper.updateByPrimaryKeySelective(record);
-    }
-
-    @Transactional
-    @Override
-    public int updateByPrimaryKey(UserInfo record) {
-        return userInfoMapper.updateByPrimaryKey(record);
+    public UserInfo update(UserInfo record) {
+        record.setUpdatedBy(getCurrUser());
+        record.setUpdatedDate(new Date());
+        int result = userInfoMapper.updateByPrimaryKeySelective(record);
+        if (result == 1) {
+            return selectOne(record.getId());
+        }
+        return null;
     }
 }
